@@ -7,12 +7,14 @@
     "cheerio": "1.0.0-rc.1",
     "comment-json": "^1.1.3",
     "express": "4.16.3",
-    "firebase-admin": "^5.13.1",
-    "firebase-functions": "^2.0.2",
+    "firebase-admin": "^8.7.0",
+    "firebase-functions": "^3.3.0",
     "fs-extra": "^5.0.0",
+    "google-auth-library": "^2.0.1",
+    "googleapis": "^35.0.0",
     "html-minifier": "3.5.19",
     "isomorphic-fetch": "2.2.1",
-    "marked": "^0.3.6",
+    "marked": "^0.3.9",
     "parse-link-header": "^1.0.1",
     "request": "^2.83.0",
     "request-promise-native": "^1.0.5",
@@ -24,6 +26,7 @@
   "scripts": {
     "dev": "npm run build && firebase functions:config:get > .runtimeconfig.json && firebase serve --only functions",
     "build": "npm run fmt && tsc",
+    "build:watch": "tsc -w",
     "fmt": "prettier --write src/*.ts src/test/*.ts"
   },
   "private": true,
@@ -34,7 +37,8 @@
     "@types/node": "^10.11.3",
     "@types/request": "^2.47.1",
     "@types/request-promise-native": "^1.0.15",
+    "firebase-functions-test": "^0.1.6",
     "prettier": "^1.8.2",
-    "typescript": "^3.1.1"
+    "typescript": "^3.4.5"
   }
 }
