@@ -1,4 +1,5 @@
 import Projects from "../../../../src/components/Projects"
+import { Env } from "../../../../../shared/types";
 
 export default {
   components: {
@@ -45,6 +46,11 @@ export default {
     const repo = context.params.repo || context.route.params.repo;
     const page = context.params.page || context.route.params.page;
 
-    return await Projects.load(org, repo, page);
+    let env: Env = Env.PROD;
+    if (context.route.path.indexOf("-staging") >= 0) {
+      env = Env.STAGING;
+    }
+
+    return await Projects.load(org, repo, page, env);
   }
 }
