@@ -32,10 +32,11 @@ export class Content {
     data: string,
     repoId: string,
     page: string | undefined,
-    config: ProjectConfig
+    config: ProjectConfig,
+    branch: string
   ): PageContent {
     const html: string = marked(data);
-    const sanitizedHtml = this.sanitizeHtml(repoId, page, config, html);
+    const sanitizedHtml = this.sanitizeHtml(repoId, page, config, html, branch);
     const sections = this.htmlToSections(sanitizedHtml);
 
     return sections;
@@ -44,11 +45,12 @@ export class Content {
   /**
    * Sanitize the content Html.
    */
-  sanitizeHtml(
+  private sanitizeHtml(
     repoId: string,
     page: string,
     config: ProjectConfig,
-    html: string
+    html: string,
+    branch: string
   ): string {
     // Links
     // * Links to page content files should go to our page
@@ -67,13 +69,22 @@ export class Content {
       }
     }
 
+    // TODO: Dynamic branch
     const renderedBaseUrl = urljoin(
-      this.getRenderedContentBaseUrl(repoId),
+      this.getRenderedContentBaseUrl(repoId, branch),
       pageDir
     );
-    const rawBaseUrl = urljoin(Github.getRawContentBaseUrl(repoId), pageDir);
+    const rawBaseUrl = urljoin(
+      Github.getRawContentBaseUrl(repoId, branch),
+      pageDir
+    );
 
     const $: CheerioStatic = cheerio.load(html);
+
+    // Make all code sections prettyprinted
+    $("pre > code").each((_: number, el: CheerioElement) => {
+      $(el).addClass("prettyprint");
+    });
 
     // Resolve all relative links to github
     const that = this;
@@ -109,7 +120,17 @@ export class Content {
       if (that._isRelativeLink(href)) {
         // Check if the link is to a page within the repo
         const repoRelative = path.join(pageDir, href);
-        if (config.pages && config.pages[repoRelative]) {
+        const pageKeys = config.pages
+          ? config.pages.map(page => page.path)
+          : [];
+
+        Logger.debug(
+          repoId,
+          `Relative link on page ${page}: ${href} --> ${repoRelative}`
+        );
+        el.attribs["href"] = repoRelative;
+
+        if (pageKeys.indexOf(repoRelative) >= 0) {
           Logger.debug(repoId, `Lowercasing relative link ${repoRelative}.`);
           that.lowercaseLink(el);
         } else {
@@ -133,7 +154,10 @@ export class Content {
         "gitter.im",
         "circleci.com",
         "opencollective.com",
-        "cirrus-ci.com"
+        "cirrus-ci.com",
+        "sonarcloud.io",
+        "codecov.io",
+        "release-notes.com"
       ];
 
       let isBadge = false;
@@ -162,7 +186,7 @@ export class Content {
   /**
    * Turn HTML into a sections objects.
    */
-  htmlToSections(html: string): PageContent {
+  private htmlToSections(html: string): PageContent {
     const $ = cheerio.load(html);
     const sections: PageSection[] = [];
 
@@ -202,15 +226,15 @@ export class Content {
   /**
    * Get the base github URL for a project.
    */
-  getRenderedContentBaseUrl(id: string) {
+  private getRenderedContentBaseUrl(id: string, branch: string) {
     // Parse the ID into pieces
     const idObj = Util.parseProjectId(id);
 
     // Get the URL to the root folder
     const pathPrefix = idObj.path ? idObj.path + "/" : "";
-    const url = `https://github.com/${idObj.owner}/${idObj.repo}/tree/master/${
-      pathPrefix
-    }`;
+    const url = `https://github.com/${idObj.owner}/${
+      idObj.repo
+    }/tree/${branch}/${pathPrefix}`;
 
     return url;
   }
@@ -218,7 +242,11 @@ export class Content {
   /**
    * Sanitize relative links to be absolute.
    */
-  sanitizeRelativeLink(el: CheerioElement, attrName: string, base: string) {
+  private sanitizeRelativeLink(
+    el: CheerioElement,
+    attrName: string,
+    base: string
+  ) {
     const val = el.attribs[attrName];
 
     if (val) {
@@ -234,7 +262,7 @@ export class Content {
   /**
    * Change href to lowercase.
    */
-  lowercaseLink(el: CheerioElement) {
+  private lowercaseLink(el: CheerioElement) {
     const newVal = el.attribs["href"].toLowerCase();
     el.attribs["href"] = newVal;
   }
@@ -242,7 +270,7 @@ export class Content {
   /**
    * Determine if a project is listed on firebaseopensource.com
    */
-  _isIncludedProject(org: string, repo: string) {
+  private _isIncludedProject(org: string, repo: string) {
     if (org === "firebase") {
       return true;
     }
@@ -257,7 +285,7 @@ export class Content {
   /**
    * Determine if a link is to github.com
    */
-  _isGithubLink(href: string) {
+  private _isGithubLink(href: string) {
     const hrefUrl = url.parse(href);
     return (
       hrefUrl.hostname && hrefUrl.pathname && href.indexOf("github.com") >= 0
@@ -267,7 +295,7 @@ export class Content {
   /**
    * Determine if a link is relative.
    */
-  _isRelativeLink(href: string) {
+  private _isRelativeLink(href: string) {
     const hrefUrl = url.parse(href);
 
     // Relative link has a pathname but not a host
