@@ -15,10 +15,11 @@
  */
 import { Project } from "./project";
 import * as functions from "firebase-functions";
+import { Cron } from "./cron";
+import { Env, GetParams } from "../../shared/types";
+import { Util } from "../../shared/util";
 
 const PubSub = require("@google-cloud/pubsub");
-
-const project = new Project();
 
 const pubsubClient = new PubSub({
   projectId: process.env.GCLOUD_PROJECT
@@ -28,6 +29,47 @@ const RUNTIME_OPTS = {
   timeoutSeconds: 540,
   memory: "2GB" as "2GB"
 };
+
+const DEFAULT_PARAMS: GetParams = {
+  env: Env.PROD,
+  branch: "master"
+};
+
+// TODO: This should be created per-function with
+//       dynamic params.
+const project = new Project(DEFAULT_PARAMS);
+
+/**
+ * Stage a project.
+ *
+ * Params: org, repo, branch
+ */
+exports.stageProject = functions
+  .runWith(RUNTIME_OPTS)
+  .https.onRequest(async (request, response) => {
+    const org = request.param("org");
+    const repo = request.param("repo");
+    const branch = request.param("branch") || "master";
+    console.log(`stageProject(${org}, ${repo}, ${branch})`);
+
+    const p = new Project({
+      env: Env.STAGING,
+      branch
+    });
+
+    const id = Util.normalizeId(`${org}::${repo}`);
+    try {
+      await p.recursiveStoreProject(id);
+      response
+        .status(200)
+        .send(
+          `Visit https://firebaseopensource.com/projects-staging/${org}/${repo}`
+        );
+    } catch (e) {
+      console.warn(e);
+      response.status(500).send(`Failed to stage project: ${e}.\n`);
+    }
+  });
 
 /**
  * Get the config and content for a single project and its subprojects.
@@ -49,16 +91,16 @@ exports.getProject = functions
 exports.getProjectWebhook = functions
   .runWith(RUNTIME_OPTS)
   .https.onRequest(async (request, response) => {
-    const id = request.param("id");
-    project
-      .recursiveStoreProject(id)
-      .then(() => {
-        response.status(200).send(`Stored project ${id}.\n`);
-      })
-      .catch(e => {
-        console.warn(e);
-        response.status(500).send(`Failed to store project ${id}: ${e}.\n`);
-      });
+    // TODO: This should just send the PubSub
+    const id = request.query["id"];
+
+    try {
+      await project.recursiveStoreProject(id);
+      response.status(200).send(`Stored project ${id}.\n`);
+    } catch (e) {
+      console.warn(e);
+      response.status(500).send(`Failed to store project ${id}: ${e}.\n`);
+    }
   });
 
 /**
@@ -84,5 +126,17 @@ exports.getAllProjects = functions
     } catch (e) {
       console.warn("Error:", e);
       response.status(500).send("Failed to store all projects.\n");
+    }
+  });
+
+exports.cloudBuild = functions
+  .runWith(RUNTIME_OPTS)
+  .https.onRequest(async (request, response) => {
+    try {
+      await Cron.build();
+      response.status(200).send("Build started!");
+    } catch (e) {
+      console.warn("Error: ", e);
+      response.status(500).send("Failed to start build.");
     }
   });
