@@ -18,37 +18,45 @@ import { Content } from "./content";
 import { Github } from "./github";
 import { Logger } from "./logger";
 import { Util } from "../../shared/util";
-import { ProjectConfig, PageSection, ProjectPage } from "../../shared/types";
+import {
+  ProjectConfig,
+  PageSection,
+  ProjectPage,
+  Env,
+  GetParams,
+  PageConfig
+} from "../../shared/types";
 
 import * as admin from "firebase-admin";
-import * as functions from "firebase-functions";
 
 const cjson = require("comment-json");
 
-const content = new Content();
-const github = new Github(Config.get("github.token"));
-
-// Initialize the Admin SDK with the right credentials for the environment
 try {
-  admin.initializeApp(functions.config().firebase);
+  admin.initializeApp();
 } catch (e) {
-  admin.initializeApp({
-    credential: admin.credential.applicationDefault()
-  });
+  Logger.error("initialize", "Could not initialize firebase-admin", e);
 }
 
 const db = admin.firestore();
 db.settings({ timestampsInSnapshots: true });
 
 export class Project {
+  params: GetParams;
+  github: Github;
+
+  constructor(params: GetParams) {
+    this.params = params;
+    this.github = new Github(Config.get("github.token"), params.branch);
+  }
+
   /**
    * Get a list of all project IDs, useful for fan-out project storing.
    */
-  listAllProjectIds = async function() {
+  async listAllProjectIds() {
     // Loads ADDITIONAL_PROJECTS
     await Config.loadGlobalConfig();
 
-    const repos = await github.listAllRepos("firebase");
+    const repos = await this.github.listAllRepos("firebase");
     const ids = repos.map(repo => {
       return Util.pathToSlug(repo);
     });
@@ -60,29 +68,29 @@ export class Project {
     );
     const allIds = ids.concat(Config.ADDITIONAL_PROJECTS);
     return allIds;
-  };
+  }
 
   /**
    * Store a project and all of its pages.
    */
-  recursiveStoreProject(id: string) {
+  async recursiveStoreProject(id: string): Promise<any> {
     Logger.debug(id, "recursiveStoreProject()");
 
-    return Config.loadGlobalConfig()
-      .then(() => {
-        return this.getProjectConfig(id);
-      })
-      .then(config => {
-        // Store this project's config and content
-        const storeConfig = this.storeProjectConfig(id, config);
-        const storeContent = this.storeProjectContent(id, config);
+    try {
+      await Config.loadGlobalConfig();
+      const config = await this.getProjectConfig(id);
 
-        // Wait for both to complete then pass on config
-        return Promise.all([storeConfig, storeContent]);
-      })
-      .catch(e => {
-        Logger.error(id, "recursiveStoreProject failed", e);
-      });
+      // Store this project's config and content
+      await this.storeProjectConfig(id, config);
+      await this.storeProjectContent(id, config);
+
+      // Only store releases in production
+      if (this.params.env == Env.PROD) {
+        await this.storeProjectReleases(id);
+      }
+    } catch (e) {
+      Logger.error(id, "recursiveStoreProject: ", e);
+    }
   }
 
   /**
@@ -114,14 +122,14 @@ export class Project {
    * Get the configuration for a project at a certain path.
    */
   getProjectConfig(id: string) {
-    const url = Github.getConfigUrl(id);
     const idParsed = Util.parseProjectId(id);
 
-    return this.checkConfigExists(id)
+    return this.github
+      .hasProjectConfig(id)
       .then((exists: boolean) => {
         if (exists) {
           // Config exists, fetch and parse it
-          return github.getContent(url).then(data => {
+          return this.github.getProjectConfig(id).then(data => {
             // Parse and remove comments
             const contents = data.toString();
             return cjson.parse(contents, null, true);
@@ -129,7 +137,7 @@ export class Project {
         } else {
           // If the project has no config, make one up
           Logger.debug(id, "WARN: Using default config.");
-          return github
+          return this.github
             .getRepoReadmeFile(idParsed.owner, idParsed.repo)
             .then((readme: string) => {
               Logger.debug(id, `README: ${readme}`);
@@ -154,9 +162,34 @@ export class Project {
           config.blacklist = false;
         }
 
+        // Normalize the formatting of "pages"
+        if (config.pages) {
+          const normalizedPages: PageConfig[] = [];
+
+          if (Array.isArray(config.pages)) {
+            // In v0 pages was just an array of paths
+            for (const path of config.pages) {
+              normalizedPages.push({
+                path
+              });
+            }
+          } else {
+            // In v1 it is a map of { path: title }
+            Object.keys(config.pages).forEach((path: string) => {
+              const name = config.pages[path];
+              normalizedPages.push({
+                name,
+                path
+              });
+            });
+          }
+
+          config.pages = normalizedPages;
+        }
+
         // Merge the config with repo metadata like stars
         // and updated time.
-        return github
+        return this.github
           .getRepoMetadata(idParsed.owner, idParsed.repo)
           .then(meta => {
             if (meta.description) {
@@ -176,14 +209,6 @@ export class Project {
   }
 
   /**
-   * Check if the config exists for a given project.
-   */
-  checkConfigExists(id: string) {
-    const url = Github.getConfigUrl(id);
-    return github.pageExists(url);
-  }
-
-  /**
    * Fetch a project config and put it into Firestore.
    */
   storeProjectConfig(id: string, config: ProjectConfig) {
@@ -193,21 +218,39 @@ export class Project {
     // Add server timestamp
     data.last_fetched = admin.firestore.FieldValue.serverTimestamp();
 
-    Logger.debug(id, `Storing at /configs/${docId}`);
-    Logger.debug(id, "Config: " + JSON.stringify(config));
-    const configProm = db
-      .collection("configs")
-      .doc(docId)
-      .set(data);
+    // Get path to the config document in the database
+    const configPath = Util.configPath(docId, this.params.env);
 
-    return configProm;
+    Logger.debug(id, `Storing at ${configPath}`);
+    Logger.debug(id, "Config: " + JSON.stringify(config));
+
+    return db.doc(configPath).set(data);
+  }
+
+  async storeProjectReleases(id: string): Promise<any> {
+    const { owner, repo } = Util.parseProjectId(id);
+    const releases = await this.github.getRepoReleases(owner, repo);
+    if (!releases || releases.length == 0) {
+      Logger.debug(id, `No releases for ${id}`);
+      return;
+    }
+
+    const batch = db.batch();
+    for (const release of releases) {
+      const releaseKey = `${id}::${release.tag_name}`;
+      const ref = db.collection("releases").doc(releaseKey);
+      batch.set(ref, release);
+    }
+
+    return batch.commit();
   }
 
   /**
    * Fetch a project's content and put it into Firestore.
    */
   async storeProjectContent(id: string, config: ProjectConfig): Promise<any> {
-    const contentRef = db.collection("content").doc(Util.normalizeId(id));
+    const contentPath = Util.contentPath(id, this.params.env);
+    const contentRef = db.doc(contentPath);
 
     const sections = await this.getProjectContent(id, config);
 
@@ -237,12 +280,17 @@ export class Project {
    * Fetch a project's content.
    */
   getProjectContent(id: string, config: ProjectConfig) {
-    const url = Github.getContentUrl(id, config);
-
-    return github
-      .getContent(url)
+    const content = new Content();
+    return this.github
+      .getRawProjectContent(id, config)
       .then(data => {
-        return content.processMarkdown(data, id, undefined, config);
+        return content.processMarkdown(
+          data,
+          id,
+          config.content,
+          config,
+          this.params.branch
+        );
       })
       .then(content => {
         content.sections = this.filterProjectSections(content.sections);
@@ -284,7 +332,7 @@ export class Project {
     id: string,
     config: ProjectConfig
   ): Promise<ProjectPage[]> {
-    if (!config.pages || Object.keys(config.pages).length == 0) {
+    if (!config.pages || config.pages.length == 0) {
       Logger.debug(id, `Project has no extra pages.`);
       return Promise.resolve([]);
     }
@@ -295,24 +343,41 @@ export class Project {
     const pages: ProjectPage[] = [];
 
     // Loop through pages, get content for each
-    Object.keys(config.pages).forEach(page => {
-      const pageUrl = Github.getPageUrl(id, page);
-      Logger.debug(id, `Rendering page: ${pageUrl}`);
+    const content = new Content();
+    for (const page of config.pages) {
+      const pageUrl = Github.getPageContentUrl(
+        id,
+        page.path,
+        this.params.branch
+      );
+      Logger.debug(
+        id,
+        `Rendering page name=${page.name}, path=${page.path}, url=${pageUrl}`
+      );
 
-      const pagePromise = github
-        .getContent(pageUrl)
+      const pagePromise = this.github
+        .getRawContent(pageUrl)
+        .catch(e => {
+          throw `Failed to get content for ${pageUrl}: ${JSON.stringify(e)}`;
+        })
         .then(data => {
-          return content.processMarkdown(data, id, page, config);
+          return content.processMarkdown(
+            data,
+            id,
+            page.path,
+            config,
+            this.params.branch
+          );
         })
         .then(sections => {
           pages.push({
-            name: page,
+            name: page.path,
             content: sections
           });
         });
 
       promises.push(pagePromise);
-    });
+    }
 
     return Promise.all(promises).then(() => {
       return pages;
