@@ -17,23 +17,38 @@
 // TODO: Delete this file
 //
 // To run:
-// GCLOUD_PROJECT="fir-oss" github_token="<YOUR_GITHUB_TOKEN>" node test/test.js
+// GCLOUD_PROJECT="fir-oss" github_token="<YOUR_GITHUB_TOKEN>" ts-node test/test.ts
 import { Config } from "../config";
 import { Project } from "../project";
+import { GetParams, Env } from "../../../shared/types";
 
-const project = new Project();
+const DEFAULT_PARAMS: GetParams = {
+  env: Env.STAGING,
+  branch: "master"
+};
+
+const project = new Project(DEFAULT_PARAMS);
 
 const ids = [
   "samtstern::BotTest",
   "firebase::quickstart-js",
   "firebase::quickstart-cpp",
   "firebase::firebaseui-android",
-  "firebase::firebaseui-ios",
   "firebase::angularfire",
   "firebase::functions-samples",
-  "firebase::assistant-codelab",
+
+  // Interesting because:
+  //  * Is "blacklisted" from being featured
   "firebase::androiddrawing",
-  "googlesamples::easypermissions"
+
+  // Interesting because:
+  //  * Uses a nested md file as it's "content" page.
+  "firebase::firebase-android-sdk",
+
+  // Interesting because:
+  //  * Non-Firebase
+  //  * Has "old-style" pages config (no names)
+  "invertase::react-native-firebase"
 ];
 
 Config.loadGlobalConfig().then(() => {
